+++ Server/recycleApp/recycleApp/src/main/java/com/recycle/recycleapp/repositories/RecycleCenterRepo.java
@@ -1,5 +1,6 @@
-package com.recycle.recycleapp.entity;
+package com.recycle.recycleapp.repositories;
 
+import com.recycle.recycleapp.entities.RecycleCenter;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
