+++ Server/recycleApp/recycleApp/src/main/java/com/recycle/recycleapp/repositories/WasteRepo.java
@@ -1,10 +1,8 @@
-package com.recycle.recycleapp.entity;
-
+package com.recycle.recycleapp.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WasteRepo extends JpaRepository<Waste,Long> {
-
+public interface WasteRepo extends JpaRepository<com.recycle.recycleapp.entity.Waste,Long> {
 }
